# docs/swagger.yaml
basePath: /
definitions:
  helper.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  model.BalanceUpdate:
    properties:
      balance:
        type: integer
    type: object
...
  model.GetTransactionUserResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      product:
        $ref: '#/definitions/model.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.LoginCredential:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      price:
        type: integer
      stock:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  model.TransactionCreate:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  model.TransactionCreateResponse:
    properties:
      message:
        type: string
      transaction_bill:
        description: TransactionBill represents the transaction bill information.
        properties:
          product_title:
            type: string
          quantity:
            type: integer
          total_price:
            type: integer
        type: object
    type: object
  model.TransactionHistory:
    properties:
      created_at:
        type: string
      id:
        type: integer
      product:
        $ref: '#/definitions/model.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.User:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      role:
        type: string
      transactionHistories:
        items:
          $ref: '#/definitions/model.TransactionHistory'
        type: array
      updated_at:
        type: string
    type: object
  model.UserRegister:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    type: object
  model.UserRegisterResponse:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
    type: object
host: final-project-04-production.up.railway.app
info:
  contact: {}
  description: API for an e-commerce platform
  title: Toko Belanja API
  version: "1.0"
paths:
...
  /transactions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transaction object to be created
        in: body
        name: model.TransactionHistory
        required: true
        schema:
          $ref: '#/definitions/model.TransactionCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            $ref: '#/definitions/model.TransactionCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - Bearer: []
      summary: Creates a new transaction
      tags:
      - Transaction
  /transactions/my-transactions:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transactions fetched successfully
          schema:
            items:
              $ref: '#/definitions/model.GetTransactionUserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - Bearer: []
      summary: Get User Transactions.
      tags:
      - Transaction
  /transactions/user-transactions:
    get:
      consumes:
      - application/json
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transactions fetched successfully
          schema:
            items:
              $ref: '#/definitions/model.TransactionHistory'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - Bearer: []
      summary: Get All Transactions.
      tags:
      - Transaction
  /users/login:
    post:
      consumes:
      - application/json
      description: Endpoint to register a new user
      parameters:
      - description: authenticate user
        in: body
        name: model.LoginCredential
        required: true
        schema:
          $ref: '#/definitions/model.LoginCredential'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "422":
          description: Invalid Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Authenticate a user
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: Endpoint to register a new user
      parameters:
      - description: create user
        in: body
        name: model.UserRegister
        required: true
        schema:
          $ref: '#/definitions/model.UserRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.UserRegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "422":
          description: Invalid Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      summary: Register a new user
      tags:
      - User
  /users/topup:
    patch:
      consumes:
      - application/json
      description: Endpoint to top-up user balance
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Amount to top-up
        in: body
        name: model.User
        required: true
        schema:
          $ref: '#/definitions/model.BalanceUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "422":
          description: Invalid Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - Bearer: []
      summary: Topup user balance
      tags:
      - User
swagger: "2.0"